+++ CHANGELOG.rst
@@ -10,10 +10,11 @@ Changelog
 * Update :meth:`pyvcf.VcfFrame.from_file` method to accept file-like object as input as well.
 * Add new method :meth:`pycov.CovFrame.mask_bed`.
 * Add new method :meth:`pycov.CovFrame.chr_prefix`.
-* Add new property ``contigs`` to :class:`pybed.BedFrame`.
+* Add new property ``contigs`` to :class:`pybed.BedFrame` class.
 * Add new method :meth:`pybed.BedFrame.chr_prefix`.
 * Add new methods :meth:`pybed.BedFrame.copy_meta` and :meth:`pybed.BedFrame.sort`.
 * Add new method :meth:`pybed.BedFrame.merge`.
+* Add new property ``empty`` to :class:`pyvcf.VcfFrame` class.
 
 0.22.0 (2021-09-04)
 -------------------